Add implementation of shrink for Arbitrary.
If we can shrink the underlying type, we can shrink the refined types
simply by checking which shrinks still satisfy the predicate.

This made prop_bad fail, I think it moved the (expected) error from a
place where expectFailure could handle it to a place where it couldn't.
So I rearranged that slightly.